PP-559 Use Alpine container for pay-logger
We have a new logging pipeline, see:
https://github.com/alphagov/pay-infra/pull/118

This means that we no longer need logs shipped to ElasticSearch. We may
use ElasticSearch/Kibana in the future, but we can stream logs to it
from CloudWatch Logs, rather than having to stream from the container
and deal with security policies for the ElasticSearch cluster, which is
a pain to maintain.

As we want a very simple syslog process to be run, and we generally want
*thin* containers, Alpine seemed to be a good fit. Jenkins build time
has gone from 25min to 1min30s :+1: